requirements.txt
Flask
werkzeug
Pillow
pytesseract
